databrowser
get table list from config json file and get data from database
