+++ web/model/query.go
@@ -1,27 +1,23 @@
 package gocore
 
 import (
-	"eaciit/gdrj/web/helper"
 	"encoding/json"
+	"errors"
 	"github.com/eaciit/dbox"
+	_ "github.com/eaciit/dbox/dbc/csv"
+	_ "github.com/eaciit/dbox/dbc/csvs"
+	_ "github.com/eaciit/dbox/dbc/json"
 	_ "github.com/eaciit/dbox/dbc/jsons"
-	"github.com/eaciit/orm/v1"
+	_ "github.com/eaciit/dbox/dbc/mongo"
 	"github.com/eaciit/toolkit"
+	"strconv"
 	"strings"
 )
 
-type DataBrowser struct {
-	orm.ModelBase
-	ID         string `json:"_id",bson:"_id"`
-	TableNames string
-	MetaData   []*StructInfo
-}
-
-type StructInfo struct {
-	Field        string
-	DataType     string
-	Sortable     bool
-	SimpleFilter bool
+type queryWrapper struct {
+	ci         *dbox.ConnectionInfo
+	connection dbox.IConnection
+	err        error
 }
 
 type MetaSave struct {
@@ -29,12 +25,35 @@ type MetaSave struct {
 	data    string
 }
 
-func (b *DataBrowser) TableName() string {
-	return "databrowser"
-}
+func Query(driver string, host string, other ...interface{}) *queryWrapper {
 
-func (b *DataBrowser) RecordID() interface{} {
-	return b.ID
+	wrapper := queryWrapper{}
+	wrapper.ci = &dbox.ConnectionInfo{host, "", "", "", nil}
+
+	if len(other) > 0 {
+		wrapper.ci.Database = other[0].(string)
+	}
+	if len(other) > 1 {
+		wrapper.ci.UserName = other[1].(string)
+	}
+	if len(other) > 2 {
+		wrapper.ci.Password = other[2].(string)
+	}
+	if len(other) > 3 {
+		wrapper.ci.Settings = other[3].(toolkit.M)
+	}
+
+	wrapper.connection, wrapper.err = dbox.NewConnection(driver, wrapper.ci)
+	if wrapper.err != nil {
+		return &wrapper
+	}
+
+	wrapper.err = wrapper.connection.Connect()
+	if wrapper.err != nil {
+		return &wrapper
+	}
+
+	return &wrapper
 }
 
 func ConnectToDatabase(payload toolkit.M) (int, []toolkit.M, *DataBrowser, error) {
@@ -43,7 +62,7 @@ func ConnectToDatabase(payload toolkit.M) (int, []toolkit.M, *DataBrowser, error
 	if payload.Has("haslookup") {
 		hasLookup = payload.Get("haslookup").(bool)
 	}
-	_id := toolkit.ToString(payload.Get("browserid", ""))
+	_id := toolkit.ToString(payload.Get("tablename", ""))
 	sort := payload.Get("sort")
 	search := payload.Get("search")
 	_ = search
@@ -67,10 +86,20 @@ func ConnectToDatabase(payload toolkit.M) (int, []toolkit.M, *DataBrowser, error
 	}
 
 	dataDS := new(DataBrowser)
-	err := Get(dataDS, _id)
+	if err := Get(dataDS, _id); err != nil {
+		return 0, nil, nil, err
+	}
+
+	/*filterDB := dbox.Contains("TableNames", tablename)
+	cursorDB, err := Find(new(DataBrowser), filterDB)
 	if err != nil {
 		return 0, nil, nil, err
 	}
+
+	dataDS := new(DataBrowser)
+	if err = cursorDB.Fetch(dataDS, 1, false); err != nil {
+		return 0, nil, nil, err
+	}*/
 
 	driver, ci := new(Login).GetConnectionInfo(CONF_DB_GDRJ)
 	connection, err := dbox.NewConnection(driver, ci)
@@ -82,7 +111,7 @@ func ConnectToDatabase(payload toolkit.M) (int, []toolkit.M, *DataBrowser, error
 	}
 
 	TblName.Set("from", dataDS.TableNames)
-	payload.Set("from", dataDS.TableNames)
+	payload.Set("from", dataDS.TableNames) /*skip, take and sort already push into payload*/
 
 	qcount, _ := parseQuery(connection.NewQuery(), TblName)
 	query, _ := parseQuery(connection.NewQuery(), payload)
@@ -243,7 +272,7 @@ func parseQuery(query dbox.IQuery, queryInfo toolkit.M) (dbox.IQuery, MetaSave) 
 
 			for _, each := range whereAll {
 				where, _ := toolkit.ToM(each)
-				filter := helper.FilterParse(where)
+				filter := FilterParse(where)
 				if filter != nil {
 					allFilter = append(allFilter, filter)
 				}
@@ -260,4 +289,216 @@ func parseQuery(query dbox.IQuery, queryInfo toolkit.M) (dbox.IQuery, MetaSave) 
 	}
 
 	return query, metaSave
+}
+
+func (c *queryWrapper) CheckIfConnected() error {
+	return c.err
+}
+
+func (c *queryWrapper) Connect() (dbox.IConnection, error) {
+	if c.err != nil {
+		return nil, c.err
+	}
+
+	return c.connection, nil
+}
+
+func (c *queryWrapper) SelectOne(tableName string, clause ...*dbox.Filter) (toolkit.M, error) {
+	if c.err != nil {
+		return nil, c.err
+	}
+
+	connection := c.connection
+	defer connection.Close()
+
+	query := connection.NewQuery().Select().Take(1)
+	if tableName != "" {
+		query = query.From(tableName)
+	}
+	if len(clause) > 0 {
+		query = query.Where(clause[0])
+	}
+
+	cursor, err := query.Cursor(nil)
+	if err != nil {
+		return nil, err
+	}
+	defer cursor.Close()
+
+	data := make([]toolkit.M, 0)
+	err = cursor.Fetch(&data, 0, false)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(data) == 0 {
+		return nil, errors.New("No data found")
+	}
+
+	return data[0], nil
+}
+
+func (c *queryWrapper) Delete(tableName string, clause *dbox.Filter) error {
+	if c.err != nil {
+		return c.err
+	}
+
+	connection := c.connection
+	defer connection.Close()
+
+	query := connection.NewQuery().Delete()
+	if tableName != "" {
+		query = query.From(tableName)
+	}
+
+	err := query.Where(clause).Exec(nil)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (c *queryWrapper) SelectAll(tableName string, clause ...*dbox.Filter) ([]toolkit.M, error) {
+	if c.err != nil {
+		return nil, c.err
+	}
+
+	connection := c.connection
+	defer connection.Close()
+
+	query := connection.NewQuery().Select()
+	if tableName != "" {
+		query = query.From(tableName)
+	}
+	if len(clause) > 0 {
+		query = query.Where(clause[0])
+	}
+
+	cursor, err := query.Cursor(nil)
+	if err != nil {
+		return nil, err
+	}
+	defer cursor.Close()
+
+	data := make([]toolkit.M, 0)
+	err = cursor.Fetch(&data, 0, false)
+	if err != nil {
+		return nil, err
+	}
+
+	return data, nil
+}
+
+func (c *queryWrapper) Save(tableName string, payload map[string]interface{}, clause ...*dbox.Filter) error {
+	if c.err != nil {
+		return c.err
+	}
+
+	connection := c.connection
+	defer connection.Close()
+
+	query := connection.NewQuery()
+	if tableName != "" {
+		query = query.From(tableName)
+	}
+
+	if len(clause) == 0 {
+		err := query.Insert().Exec(toolkit.M{"data": payload})
+		if err != nil {
+			return err
+		}
+
+		return nil
+	} else {
+		err := query.Update().Where(clause[0]).Exec(toolkit.M{"data": payload})
+		if err != nil {
+			return err
+		}
+
+		return nil
+	}
+
+	return errors.New("nothing changes")
+}
+
+func FilterParse(where toolkit.M) *dbox.Filter {
+	field := where.Get("field", "").(string)
+	value := toolkit.Sprintf("%v", where["value"])
+
+	if key := where.Get("key", "").(string); key == "Eq" {
+		valueInt, errv := strconv.Atoi(toolkit.Sprintf("%v", where["value"]))
+		if errv == nil {
+			return dbox.Eq(field, valueInt)
+		} else {
+			return dbox.Eq(field, value)
+		}
+	} else if key == "Ne" {
+		valueInt, errv := strconv.Atoi(toolkit.Sprintf("%v", where["value"]))
+		if errv == nil {
+			return dbox.Ne(field, valueInt)
+		} else {
+			return dbox.Ne(field, value)
+		}
+	} else if key == "Lt" {
+		valueInt, errv := strconv.Atoi(toolkit.Sprintf("%v", where["value"]))
+		if errv == nil {
+			return dbox.Lt(field, valueInt)
+		} else {
+			return dbox.Lt(field, value)
+		}
+	} else if key == "Lte" {
+		valueInt, errv := strconv.Atoi(toolkit.Sprintf("%v", where["value"]))
+		if errv == nil {
+			return dbox.Lte(field, valueInt)
+		} else {
+			return dbox.Lte(field, value)
+		}
+	} else if key == "Gt" {
+		valueInt, errv := strconv.Atoi(toolkit.Sprintf("%v", where["value"]))
+		if errv == nil {
+			return dbox.Gt(field, valueInt)
+		} else {
+			return dbox.Gt(field, value)
+		}
+	} else if key == "Gte" {
+		valueInt, errv := strconv.Atoi(toolkit.Sprintf("%v", where["value"]))
+		if errv == nil {
+			return dbox.Gte(field, valueInt)
+		} else {
+			return dbox.Gte(field, value)
+		}
+	} else if key == "In" {
+		valueArray := []interface{}{}
+		for _, e := range strings.Split(value, ",") {
+			valueArray = append(valueArray, strings.Trim(e, ""))
+		}
+		return dbox.In(field, valueArray...)
+	} else if key == "Nin" {
+		valueArray := []interface{}{}
+		for _, e := range strings.Split(value, ",") {
+			valueArray = append(valueArray, strings.Trim(e, ""))
+		}
+		return dbox.Nin(field, valueArray...)
+	} else if key == "Contains" {
+		return dbox.Contains(field, value)
+	} else if key == "Or" {
+		subs := where.Get("value", []interface{}{}).([]interface{})
+		filtersToMerge := []*dbox.Filter{}
+		for _, eachSub := range subs {
+			eachWhere, _ := toolkit.ToM(eachSub)
+			filtersToMerge = append(filtersToMerge, FilterParse(eachWhere))
+		}
+		return dbox.Or(filtersToMerge...)
+	} else if key == "And" {
+		subs := where.Get("value", []interface{}{}).([]interface{})
+		filtersToMerge := []*dbox.Filter{}
+		for _, eachSub := range subs {
+			eachWhere, _ := toolkit.ToM(eachSub)
+			filtersToMerge = append(filtersToMerge, FilterParse(eachWhere))
+		}
+		return dbox.And(filtersToMerge...)
+	}
+
+	return nil
 }
